# build/resources/main/openapi.yml
openapi: 3.0.3
info:
  title: MDM (Monolithic Document Management) API
  description: |
    모놀리식 문서 관리 시스템 API
    
    현재 구현된 기능:
    - JWT 기반 인증 시스템 (회원가입, 로그인, 비밀번호 재설정)
    - 문서 수집 시스템 (단일 등록, 벌크 등록, 재등록)
    - 레포지토리 분석 및 문서 추출
    - 승인 워크플로우 (승인 대기 중)
    
    계획된 기능:
    - 승인 관리 시스템
    - 카탈로그 관리
    - 문서 검색 및 RAG 채팅
    - 파일 업로드 및 관리
  version: 1.0.0
  contact:
    name: MDM Team
    email: mdm@company.com

servers:
  - url: http://localhost:8080/api
    description: Development server

security:
  - bearerAuth: []

paths:
  # Authentication - 구현 완료
  /auth/signup:
    post:
      tags: [Authentication]
      summary: 사용자 회원가입
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: 회원가입 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: 이미 존재하는 이메일
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags: [Authentication]
      summary: 사용자 로그인
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/password/forgot:
    post:
      tags: [Authentication]
      summary: 비밀번호 재설정 요청
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        '200':
          description: 비밀번호 재설정 이메일 발송
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "비밀번호 재설정 이메일이 발송되었습니다."
        '404':
          description: 사용자를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Ingest - 구현 완료
  /ingest/single:
    post:
      tags: [Ingest]
      summary: 개별 등록 (레포 URL 또는 파일)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SingleIngestRequest'
      responses:
        '200':
          description: 레포지토리 미리보기 (유사 항목 포함)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestPreviewResponse'
        '201':
          description: 파일 기반 등록 요청 생성 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestRequestResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "레포 URL 또는 파일 ID가 필요합니다."

  /ingest/bulk:
    post:
      tags: [Ingest]
      summary: 벌크 등록
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                type:
                  type: string
                  enum: [ZIP, CSV]
              required: [file, type]
      responses:
        '200':
          description: 벌크 등록 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkIngestResponse'
        '400':
          description: 처리 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /ingest/{id}/resubmit:
    post:
      tags: [Ingest]
      summary: 재등록/수정 (버전업)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResubmitRequest'
      responses:
        '200':
          description: 재등록 요청 생성
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestRequestResponse'
        '400':
          description: 처리 오류
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  # 미구현 API들 - 계획된 기능
  /approval/requests:
    get:
      tags: [Approval - Planned]
      summary: 승인 요청 목록 조회 (미구현)
      description: "계획된 기능: 관리자가 승인 대기 중인 요청들을 조회"
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED, IMPORTED]
        - name: owner
          in: query
          schema:
            type: string
      responses:
        '501':
          description: 미구현 기능

  /approval/requests/{id}/decide:
    post:
      tags: [Approval - Planned]
      summary: 승인/반려 결정 (미구현)
      description: "계획된 기능: 관리자가 등록 요청을 승인하거나 반려"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '501':
          description: 미구현 기능

  /catalog/tree:
    get:
      tags: [Catalog - Planned]
      summary: 카탈로그 트리 조회 (미구현)
      description: "계획된 기능: 대분류-중분류-소분류 카탈로그 구조 조회"
      responses:
        '501':
          description: 미구현 기능

  /documents/{id}:
    get:
      tags: [Document - Planned]
      summary: 문서 상세 조회 (미구현)
      description: "계획된 기능: 등록된 문서의 상세 정보 조회"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '501':
          description: 미구현 기능

  /search/similar:
    post:
      tags: [Search - Planned]
      summary: 유사 문서 검색 (미구현)
      description: "계획된 기능: 벡터 검색을 통한 유사 문서 찾기"
      responses:
        '501':
          description: 미구현 기능

  /search/query:
    post:
      tags: [Search - Planned]
      summary: 하이브리드 검색 (미구현)
      description: "계획된 기능: 벡터 + 키워드 하이브리드 검색"
      responses:
        '501':
          description: 미구현 기능

  /chat/query:
    post:
      tags: [Chat - Planned]
      summary: RAG 챗봇 질의 (미구현)
      description: "계획된 기능: 등록된 문서 기반 RAG 채팅"
      responses:
        '501':
          description: 미구현 기능

  /files/presign:
    post:
      tags: [File - Planned]
      summary: 파일 업로드 URL 생성 (미구현)
      description: "계획된 기능: S3 presigned URL 생성"
      responses:
        '501':
          description: 미구현 기능

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    # Authentication Schemas - 구현 완료
    SignupRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          example: "password123"
        name:
          type: string
          example: "홍길동"
      required: [email, password, name]

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          example: "password123"
      required: [email, password]

    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: '#/components/schemas/User'

    User:
      type: object
      properties:
        id:
          type: string
          example: "507f1f77bcf86cd799439011"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "홍길동"
        roles:
          type: array
          items:
            type: string
            enum: [USER, APPROVER, ADMIN]
          example: ["USER"]
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time

    # Ingest Schemas - 구현 완료
    SingleIngestRequest:
      type: object
      properties:
        repoUrl:
          type: string
          format: uri
          example: "https://github.com/user/repository"
          description: "GitHub 레포지토리 URL (레포지토리 등록 시)"
        accessToken:
          type: string
          example: "ghp_xxxxxxxxxxxxxxxxxxxx"
          description: "GitHub 액세스 토큰 (Private 레포지토리 접근 시)"
        fileIds:
          type: array
          items:
            type: string
          example: ["file_id_1", "file_id_2"]
          description: "업로드된 파일 ID 목록 (파일 등록 시)"
        purpose:
          type: string
          example: "프로젝트 계획서"
          description: "문서/레포지토리의 목적"
        tags:
          type: array
          items:
            type: string
          example: ["web", "react", "typescript"]
          description: "태그 목록"

    IngestPreviewResponse:
      type: object
      properties:
        extractedText:
          type: string
          example: "이 프로젝트는 React를 사용한 웹 애플리케이션입니다..."
          description: "추출된 텍스트 (README, 코드 분석 결과)"
        proposedCategory:
          $ref: '#/components/schemas/Category'
        proposedPurpose:
          type: string
          example: "React 웹 애플리케이션 개발 프로젝트"
        similarCandidates:
          type: array
          items:
            $ref: '#/components/schemas/SimilarCandidate'

    IngestRequestResponse:
      type: object
      properties:
        id:
          type: string
          example: "507f1f77bcf86cd799439011"
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED, IMPORTED]
          example: "PENDING"
        requestedAt:
          type: string
          format: date-time

    BulkIngestResponse:
      type: object
      properties:
        totalProcessed:
          type: integer
          example: 50
        successCount:
          type: integer
          example: 45
        errorCount:
          type: integer
          example: 5
        errors:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
                example: 3
              error:
                type: string
                example: "잘못된 URL 형식"

    ResubmitRequest:
      type: object
      properties:
        purpose:
          type: string
          example: "수정된 프로젝트 목적"
        tags:
          type: array
          items:
            type: string
          example: ["updated", "version2"]
        reason:
          type: string
          example: "요구사항 변경으로 인한 재등록"

    # Common Schemas
    Category:
      type: object
      properties:
        majorCode:
          type: string
          example: "A"
        majorName:
          type: string
          example: "웹 개발"
        midCode:
          type: string
          example: "A01"
        midName:
          type: string
          example: "프론트엔드"
        subCode:
          type: string
          example: "A0101"
        subName:
          type: string
          example: "React"

    SimilarCandidate:
      type: object
      properties:
        docId:
          type: string
          example: "507f1f77bcf86cd799439011"
        score:
          type: number
          format: float
          example: 0.85
          description: "유사도 점수 (0-1)"
        serial:
          type: string
          example: "A0101-001"
          description: "문서 일련번호"
        purpose:
          type: string
          example: "React 대시보드 프로젝트"
        snippet:
          type: string
          example: "React를 사용한 관리자 대시보드 개발..."

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "CONFLICT"
        message:
          type: string
          example: "이미 존재하는 이메일입니다."
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: "/auth/signup"

  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Conflict:
      description: 리소스 충돌
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    NotImplemented:
      description: 미구현 기능
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "NOT_IMPLEMENTED"
              message:
                type: string
                example: "이 기능은 아직 구현되지 않았습니다."