# src/main/java/com/company/app/auth/dto/AuthResponse.java
package com.company.app.auth.dto;

import com.company.app.auth.entity.User;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;
import java.util.List;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class AuthResponse {
    private String accessToken;
    private String refreshToken;
    private UserDto user;
    
    @Data
    @NoArgsConstructor
    @AllArgsConstructor
    public static class UserDto {
        private String id;
        private String email;
        private String name;
        private List<User.Role> roles;
        private LocalDateTime createdAt;
        private LocalDateTime lastLoginAt;
    }
}